+++ CHANGES.md
@@ -2,8 +2,11 @@
 # CHANGES TO RELEASE
 
 1.0.4
+
 - NEW: sct_crop: function to quickly crop an image.
-- REF: sct_estimate_MAP_tracts.py now called sct_estimate_tracts.py. New functionalities added (maximum likelihood estimation and tract grouping).
+- NEW: sct_extract_metric.py (replaces the old sct_estimate_MAP_tracts.py). New functionalities added (maximum likelihood estimation and tract grouping). More flexible with label files.
+- NEW: sct_convert_mnc2nii.py
+- OPT: new version of the white matter atlas: more accurate, deformation accounting for internal structure (use BSplineSyN instead of SyN).
 
 1.0.3 (2014-07-30)